LRNG - add atomic DRNG implementation
The atomic DRNG implementation supports the in-kernel use cases which
request random numbers in atomic contexts. It uses the ChaCha20 DRNG
which is a code base that does not sleep and it provides an interface to
callers that does not sleep. This code will only be compiled when the
LRNG enables the in-kernel drop-in replacement APIs for the existing
random.c code base.

Change-Id: I1ce6868489a1661141316f7e2a7987d7d04b4114
Signed-off-by: Stephan Mueller <smueller@chronox.de>